+++ sqlite/read_test.go
@@ -22,12 +22,14 @@ func TestTableRead(t *testing.T) {
 	sql.CreateTable(testtablename, TableTest{})
 	defer os.Remove(testdbname)
 
+	ckdata := []TableTest{}
+
 	wdata := TableTest{Id: 1, Str: "data", I: 500}
-	sql.Add(testtablename, &wdata)
-	wdata1 := TableTest{Id: 1, Str: "data1", I: 45500}
-	sql.Add(testtablename, &wdata1)
-	wdata2 := TableTest{Id: 1, Str: "data2", I: 5500}
-	sql.Add(testtablename, &wdata2)
+	for i := 0; i < 20; i++ {
+		wdata.I += i
+		sql.Add(testtablename, &wdata)
+		ckdata = append(ckdata, wdata)
+	}
 
 	t.Log("-----------Read data ---------------")
 	rdata := []TableTest{}
@@ -36,7 +38,6 @@ func TestTableRead(t *testing.T) {
 		t.Errorf("read err :%v", err.Error())
 		t.FailNow()
 	}
-	ckdata := []TableTest{wdata, wdata1, wdata2}
 	if len(ckdata) != len(rdata) {
 		t.Errorf("read count ng %v=%v", len(ckdata), len(rdata))
 		t.FailNow()
@@ -50,25 +51,31 @@ func TestTableRead(t *testing.T) {
 			t.Errorf("read ng for str %v=%v", tmp.Str, tmp2.Str)
 		}
 	}
+	t.Log(rdata)
+	t.Log("read all OK")
 	t.Log("-----------Read data Serch by ID---------------")
 	rdata1 := []TableTest{}
-	err1 := sql.Read(testtablename, TableTest{}, &rdata1, map[string]string{"id": "1"}, AND)
+	err1 := sql.Read(testtablename, TableTest{}, &rdata1, map[string]string{"id": "2"}, AND)
 	if err1 != nil {
 		t.Errorf("read err :%v", err1.Error())
 	}
-	if len(rdata) != 1 {
-		t.Errorf("read count ng %v=%v", 1, len(rdata))
+	if len(rdata1) != 1 {
+		t.Errorf("read count ng %v=%v", 1, len(rdata1))
 		t.FailNow()
 	}
-	for i, tmp := range rdata {
-		tmp2 := ckdata[i]
-		if tmp.I != tmp2.I {
-			t.Errorf("read ng for i %v=%v", tmp.I, tmp2.I)
-		}
-		if tmp.Str != tmp2.Str {
-			t.Errorf("read ng for str %v=%v", tmp.Str, tmp2.Str)
+	for _, tmp := range rdata1 {
+		for _, tmp2 := range ckdata {
+			if tmp2.Id == tmp.Id {
+				if tmp.I != tmp2.I {
+					t.Errorf("read ng for i %v=%v", tmp.I, tmp2.I)
+				}
+				if tmp.Str != tmp2.Str {
+					t.Errorf("read ng for str %v=%v", tmp.Str, tmp2.Str)
+				}
+			}
 		}
 	}
+	t.Log("ID serch OK")
 	t.Log("table data read OK")
 
 }